rename rasta_lib.h to rasta_init.h
