+++ src/c/rasta_init.c
@@ -1,13 +1,16 @@
-#include <rasta/rasta_lib.h>
+#include <rasta/rasta_init.h>
 
 #include <memory.h>
 #include <stdbool.h>
-#include <rasta/rasta.h>
 #include <rasta/rmemory.h>
 #include "transport/transport.h"
 #include "transport/events.h"
 #include "retransmission/protocol.h"
 #include "retransmission/safety_retransmission.h"
+
+// This is the time that packets are deferred for creating multi-packet messages
+// See section 5.5.10
+#define IO_INTERVAL 10000
 
 void init_connection_timeout_event(timed_event *ev, struct timed_event_data *carry_data,
                                    struct rasta_connection *connection) {
@@ -56,9 +59,13 @@ void init_connection_events(struct rasta_handle *h, struct rasta_connection *con
 #endif
 }
 
-// This is the time that packets are deferred for creating multi-packet messages
-// See section 5.5.10
-#define IO_INTERVAL 10000
+void rasta_socket(rasta_lib_configuration_t user_configuration, rasta_config_info *config, struct logger_t *logger) {
+    struct rasta_handle *handle = &user_configuration->h;
+    rasta_handle_init(handle, config, logger);
+
+    //  register redundancy layer diagnose notification handler
+    handle->mux.notifications.on_diagnostics_available = handle->notifications.on_redundancy_diagnostic_notification;
+}
 
 void rasta_lib_init_configuration(rasta_lib_configuration_t user_configuration, rasta_config_info *config, struct logger_t *logger, rasta_connection_config *connections, size_t connections_length) {
     memset(user_configuration, 0, sizeof(rasta_lib_configuration_t));
