+++ src/c/retransmission/handlers.c
@@ -2,6 +2,8 @@
 #include "safety_retransmission.h"
 #include "protocol.h"
 
+#include "../experimental/handlers.h"
+
 int handle_received_packet(struct rasta_connection *connection, struct RastaPacket *receivedPacket) {
     switch (receivedPacket->type) {
         case RASTA_TYPE_RETRDATA:
